backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
pypdf
python-docx
